2006-08-15	Joel Sherrill <joel@OARcorp.com>
	* libnetworking/kern/kern_sysctl.c: Change int to int32_t.
	* libnetworking/net/if_loop.c: Add cast to long.
	These plus a patch to machine/param.h let the h8300 build multilib.

+++ cpukit/libnetworking/net/if_loop.c
@@ -233,7 +233,7 @@ lortrequest(cmd, rt, sa)
 		 * overhead.
 		 */
 		rt->rt_rmx.rmx_recvpipe = 
-			rt->rt_rmx.rmx_sendpipe = 3 * LOMTU;
+			rt->rt_rmx.rmx_sendpipe = 3L * (long)LOMTU;
 	}
 }
